Move KernelKey into testlib/dbus and have it use StratisDbus methods
If sharing clean_up, might as well share KernelKey as well and vice-versa.
We previously thought we might split CLI and daemon tests into separate
repos, but as that hasn't happened yet, and more and more code is being
shared, might as well make the most of the path we are on.
